requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
mysql-connector-python==8.4.0
python-dotenv==1.0.1
